corewars_parser
Function parse

pub fn parse(input: &str) -> Result<Warrior>

Parse a given input string into a Result. If successful the Result will contain a Warrior, otherwise it will contain an error. In either case, one or more Warnings may be generated with the Warrior.