Crate corewars_parser

This module is used for parsing a Redcode program. It operates in multiple phases, which are found in the phase module. Each phase passes its result to the next phase.

Enums

Error

An error that occurred while parsing a warrior.

Result

Result mimics the std::result::Result type, but each variant also carries zero or more Warnings with it.

Warning

A warning that occurred while parsing a warrior.

Functions

parse

Parse a given input string into a Result. If successful the Result will contain a Warrior, otherwise it will contain an error. In either case, one or more Warnings may be generated with the Warrior.